ref: separate large text response from index.ts
